Move tests for specific indexes into the appropriate child package.
